split up more tests
